+++ numhask/src/NumHask/Algebra/Abstract/Module.hs
@@ -8,7 +8,7 @@
 {-# OPTIONS_GHC -Wall #-}
 
 -- | Algebra for Modules
-module NumHask.Algebra.Module
+module NumHask.Algebra.Abstract.Module
   ( AdditiveModule(..)
   , AdditiveGroupModule(..)
   , MultiplicativeModule(..)
@@ -19,10 +19,10 @@ module NumHask.Algebra.Module
   , TensorProduct(..)
   ) where
 
-import NumHask.Algebra.Additive
+import NumHask.Algebra.Addition
 import NumHask.Algebra.Field
 import NumHask.Algebra.Metric
-import NumHask.Algebra.Multiplicative
+import NumHask.Algebra.Multiplication
 import NumHask.Algebra.Ring
 import Data.Int (Int8, Int16, Int32, Int64)
 import Data.Word (Word, Word8, Word16, Word32, Word64)
@@ -36,7 +36,7 @@ import Prelude
 -- > (a + b) .+ c == (a .+ c) + b
 -- > a .+ zero == a
 -- > a .+ b == b +. a
-class (Additive a) =>
+class (Addition a) =>
       AdditiveModule r a where
   infixl 6 .+
   (.+) :: r a -> a -> r a
@@ -82,31 +82,6 @@ class (MultiplicativeGroup a, MultiplicativeModule r a) =>
   (./) :: r a -> a -> r a
   infixl 7 /.
   (/.) :: a -> r a -> r a
-
--- | Banach (with Norm) laws form rules around size and direction of a number, with a potential crossing into another codomain.
---
--- > a == singleton zero || normalizeL2 a *. normL2 a == a
-class (ExpField a, Normed (r a) a, MultiplicativeGroupModule r a) =>
-      Banach r a where
-  normalizeL1 :: r a -> r a
-  normalizeL1 a = a ./ normL1 a
-
-  normalizeL2 :: r a -> r a
-  normalizeL2 a = a ./ normL2 a
-
-  normalizeLp :: a -> r a -> r a
-  normalizeLp p a = a ./ normLp p a
-
--- | the inner product of a representable over a semiring
---
--- > a <.> b == b <.> a
--- > a <.> (b +c) == a <.> b + a <.> c
--- > a <.> (s *. b + c) == s * (a <.> b) + a <.> c
--- (s0 *. a) <.> (s1 *. b) == s0 * s1 * (a <.> b)
-class (Semiring a) =>
-      Hilbert r a where
-  infix 8 <.>
-  (<.>) :: r a -> r a -> a
 
 -- | tensorial type
 type family (><) (a :: k1) (b :: k2) :: *
